Moved src/Host.ts to src/types/Host.ts
