+++ src/DefaultHostValidator.ts
@@ -1,4 +1,4 @@
-import Host from './Host';
+import Host from './types/Host';
 import HostValidator from './HostValidator'
 
 let validate: HostValidator;
