+++ src/HostController.ts
@@ -1,6 +1,6 @@
 import { IncomingMessage, ServerResponse } from "http"
 import { HostRepository } from "./HostRepository"
-import Host, { HostPage, HostSaveResult, SaveStatus } from './Host'
+import Host, { HostPage, HostSaveResult, SaveStatus } from './types/Host'
 import validate from './DefaultHostValidator'
 import LogService from "./services/LogService";
