+++ src/HostValidator.ts
@@ -1,4 +1,4 @@
-import Host from './Host';
+import Host from './types/Host';
 
 export default interface HostValidator {
     (host: Host): Promise<Host>
